Migrate from uint*_t types to u* types
ex: uint8_t -> u8

+++ src/ArchiveFormats/ArchiveFormat.h
@@ -1,13 +1,13 @@
 #pragma once
 
-#include <stdint.h>
-
 #include "ExeFile.h"
 #include "Entry.h"
 #include <cstring>
 #include <unordered_map>
 #include <algorithm>
-#include "../util/Logger.h" // IWYU pragma: keep
+#include <util/Logger.h>
+#include <util/int.h>
+#include "zero_templates.h"
 
 template <typename T = std::string>
 bool VectorHas(std::vector<T> vec, T item) {
@@ -16,7 +16,7 @@ bool VectorHas(std::vector<T> vec, T item) {
 
 class ArchiveBase {
     public:
-        virtual const char* OpenStream(const Entry *entry, uint8_t *buffer) = 0;
+        virtual const char* OpenStream(const Entry *entry, u8 *buffer) = 0;
         virtual std::unordered_map<std::string, Entry*> GetEntries() = 0;
         virtual ~ArchiveBase() = default;
 };
@@ -25,31 +25,31 @@ class ArchiveFormat {
     public:
         std::string tag = "?????";
         std::string description = "????? Resource Archive";
-        uint32_t sig = 0x00000000;
+        u32 sig = 0x00000000;
 
         size_t buffer_position = 0;
 
         template<typename T>
-        T Read(uint8_t* buffer, size_t offset) const {
+        T Read(u8 *buffer, usize offset) const {
             return *based_pointer<T>(buffer, offset);
         }
         template<typename T>
-        T Read(uint8_t* buffer) {
+        T Read(u8 *buffer) {
             T read = *based_pointer<T>(buffer, buffer_position);
             buffer_position += sizeof(T);
             return read;
         }
 
-        void Read(void *dest, uint8_t *src, size_t size) {
+        void Read(void *dest, u8 *src, usize size) {
             memcpy(dest, src + buffer_position, size);
             buffer_position += size;
         }
 
-        void Seek(size_t new_position) {
+        void Seek(usize new_position) {
             buffer_position = new_position;
         }
 
-        void Advance(size_t amount) {
+        void Advance(usize amount) {
             buffer_position += amount;
         }
 
@@ -58,18 +58,18 @@ class ArchiveFormat {
         }
 
         template<typename T>
-        T ReadMagic(uint8_t *buffer) const {
+        T ReadMagic(u8 *buffer) const {
             return *based_pointer<T>(buffer, 0);
         }
 
-        std::string ReadString(uint8_t *buffer, uint64_t offset) {
+        std::string ReadString(u8 *buffer, u64 offset) {
             std::string constructed = "";
             constructed.append(based_pointer<char>(buffer, offset));
 
             return constructed;
         }
 
-        std::string ReadStringAndAdvance(uint8_t *buffer, uint64_t offset) {
+        std::string ReadStringAndAdvance(u8 *buffer, u64 offset) {
             std::string constructed = "";
             constructed.append(based_pointer<char>(buffer, offset));
             buffer_position += constructed.length();
@@ -77,7 +77,7 @@ class ArchiveFormat {
             return constructed;
         }
 
-        std::string ReadStringAndAdvance(uint8_t *buffer, uint64_t offset, int length) {
+        std::string ReadStringAndAdvance(u8 *buffer, u64 offset, usize length) {
             std::string constructed = "";
             constructed.append(based_pointer<char>(buffer, offset), length);
             buffer_position += length;
@@ -85,23 +85,23 @@ class ArchiveFormat {
             return constructed;
         }
 
-        std::string ReadStringWithLength(const uint8_t* buffer, int length) {
+        std::string ReadStringWithLength(const u8* buffer, usize length) {
             return std::string((const char*)(buffer), length);
         }
 
-        ExeFile* ConvertToExeFile(uint8_t *buffer) {
+        ExeFile* ConvertToExeFile(u8 *buffer) {
             return new ExeFile(buffer);
         }
 
         // Taken directly from GARbro GameRes->ArchiveFormat.cs
-        bool IsSaneFileCount(uint32_t file_count) {
+        bool IsSaneFileCount(u32 file_count) {
             return file_count > 0 && file_count < 0x40000;
         }
 
         virtual ~ArchiveFormat() = default;
 
-        virtual bool CanHandleFile(uint8_t *buffer, uint64_t size, const std::string &ext) const = 0;
-        virtual ArchiveBase* TryOpen(uint8_t *buffer, uint64_t size, std::string file_name) = 0;
+        virtual bool CanHandleFile(u8 *buffer, u64 size, const std::string &ext) const = 0;
+        virtual ArchiveBase* TryOpen(u8 *buffer, u64 size, std::string file_name) = 0;
         virtual std::string GetTag() const {
             return this->tag;
         };
